requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.3.0
astroid==3.1.0
async-timeout==4.0.3
attrs==23.2.0
black==24.4.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
coverage==7.5.0
cycler==0.12.1
dataclasses-json==0.6.4
datamodel-code-generator==0.25.5
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
genbadge==1.1.1
genson==1.2.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpretty==1.1.4
httpx==0.27.0
identify==2.5.35
idna==3.7
inflect==5.6.2
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-experimental==0.0.57
langchain-mongodb==0.1.3
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.39
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
openai==1.16.1
orjson==3.10.0
packaging==23.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
prometheus_client==0.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pylint==3.1.0
pymongo==4.6.3
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-snapshot==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
responses==0.25.0
respx==0.21.1
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tavily-python==0.3.3
tenacity==8.2.3
testcontainers==4.4.0
tiktoken==0.6.0
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
wrapt==1.16.0
yarl==1.9.4
